refactor(use-token): include token set
